requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
PyYAML>=6.0
requests>=2.28.0
pytz>=2023.3
